Ultima update
Deleção implementada :)

+++ src/main/java/br/edu/infnet/rdsdemo/controller/DeletarController.java
@@ -8,6 +8,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
+import br.edu.infnet.rdsdemo.model.Pessoa;
+import br.edu.infnet.rdsdemo.service.AmazonService;
 import br.edu.infnet.rdsdemo.service.PessoaService;
 
 @Controller
@@ -17,14 +19,22 @@ public class DeletarController {
     @Autowired
     private PessoaService pessoaService;
 
+    @Autowired
+    public AmazonService amazonService;
+
     @GetMapping
     public String deletar() {
         return "redirect:/pessoas";
     }
 
     @GetMapping("/{id}")
-    public String deletarPessoa(@PathVariable("id") Long id, Model model) {
-        pessoaService.deletePessoaById(id);
+    public String deletarPessoa(@PathVariable("id") Long id) {
+        Pessoa pessoa = pessoaService.getPessoaById(id);
+        if(pessoa != null) {
+            amazonService.deleteFileFromS3Bucket(pessoa.getImagemPerfil());
+            pessoaService.deletePessoaById(id);
+        }
+        
         return "redirect:/pessoas";
     }
